WEB: Fix date for news on the archive and news by date pages
The news page that limits news to the last few news was properly handling file names with a character after the date (used when we
have several news for the same day).  However the pages for the
news archive or for the news for a date were not removing that
additional character and the link for the header of the news had
this incorrect date with the additional character.