WEB: Fix sorting of recent news when we have several news for the same date
The news archive page and news for a date page where using natural sorting for
the dates, which properly handles dates with an additional alphabetic character
at the end (used when we have several news for the same day). However the
recent news code was using a numeric sort, which resulted in those dates not
being properly handled and the order being the reverse of what we want.